requirements.txt
streamlit
yfinance
alpaca-py
alpaca-trade-api
pandas
numpy
requests
PyYAML
statsmodels
asyncpg